refactor : PostingDetailResponse 패키지 이동
